fix: (:dac channels desync) At last i have synced outputs. The issue was at the way dac was configured to start conversions.
The problem was i thought i dont have a way to serve both DAC channels
with one synced timer and when i tried i was giving priorities to the
channels by default. The number of the channel was triggering first. ST
SoC's have a mode to serve a 32 bit data word to a special register
(DHR12RD). Using the aformentioned register lets you put both channels
in one DMA transfer making the channels in sync.

See also RM-0444 pg. 427.

+++ include/dma.h
@@ -2,7 +2,7 @@
 #include "stm32g071xx.h"
 #include "stm32g0xx_ll_dma.h"
 struct dma {
-    const uint16_t      *data;
+    const uint32_t      *data;
     uint32_t             channel;
     DMA_TypeDef         *dmax;
     LL_DMA_InitTypeDef   dmax_settings;
