+++ src/nco.c
@@ -70,9 +70,9 @@ static inline void mmcpy ( void* dst, const void* src, uint16_t length) {
 }
 
 __attribute__((always_inline))
-inline void update_data_buff(const volatile  uint16_t data[static 128],
-                             uint16_t       buffer_sector[static 128],
+inline void update_data_buff(const volatile  uint32_t data[static 128],
+                             uint32_t       buffer_sector[static 128],
                              uint16_t       sector_length) {
     /* memcpy(buffer_sector, (const void* restrict) data, sizeof(uint16_t) * sector_length); */
-    mmcpy(buffer_sector, (const void*) data, sizeof(uint16_t) * sector_length);
+    mmcpy(buffer_sector, (const void*) data, sizeof(uint32_t) * sector_length);
 }
